requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
bert-extractive-summarizer==0.10.1
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
google_search_results==2.4.2
googletrans==3.0.0
grpcio==1.66.1
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.8.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.24.6
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
murmurhash==1.0.10
namex==0.0.8
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
preshed==3.0.9
protobuf==4.25.4
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rfc3986==1.5.0
rich==13.8.0
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
setuptools==74.0.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
